# backend/storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "L5 Swagger API",
        "description": "L5 Swagger API description",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "version": "1.0.0"
    },
    "host": "task-treking/public/",
    "basePath": "/api",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/index/{id}": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Get clients information",
                "description": "Returns clients data",
                "operationId": "getIndexById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Clients"
                ],
                "summary": "Update new client",
                "parameters": [
                    {
                        "name": "client",
                        "in": "body",
                        "description": "Client object that needs to be added to the store",
                        "schema": {
                            "properties": {
                                "id": {
                                    "description": "Class ClientsController",
                                    "type": "integer"
                                },
                                "name": {
                                    "description": "Class ClientsController",
                                    "type": "string"
                                },
                                "email": {
                                    "description": "Class ClientsController",
                                    "type": "string"
                                },
                                "primary_number": {
                                    "description": "Class ClientsController",
                                    "type": "string"
                                },
                                "secondary_number": {
                                    "description": "Class ClientsController",
                                    "type": "string"
                                },
                                "address": {
                                    "description": "Class ClientsController",
                                    "type": "string"
                                },
                                "zipcode": {
                                    "description": "Class ClientsController",
                                    "type": "string"
                                },
                                "city": {
                                    "description": "Class ClientsController",
                                    "type": "string"
                                },
                                "company_name": {
                                    "description": "Class ClientsController",
                                    "type": "string"
                                },
                                "vat": {
                                    "description": "Class ClientsController",
                                    "type": "string"
                                },
                                "company_type": {
                                    "description": "Class ClientsController",
                                    "type": "string"
                                },
                                "user_id": {
                                    "description": "Class ClientsController",
                                    "type": "integer"
                                },
                                "industry_id": {
                                    "description": "Class ClientsController",
                                    "type": "integer"
                                }
                            }
                        },
                        "default": [
                            {
                                "name": "Taras",
                                "email": "taras2andry@mail.ru",
                                "primary_number": "507212852",
                                "secondary_number": "507212852",
                                "address": "Kalush",
                                "zipcode": "77300",
                                "city": "Kalush",
                                "company_name": "IFNTUOG",
                                "vat": 23,
                                "company_type": "IT",
                                "user_id": 1,
                                "industry_id": 1
                            }
                        ]
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Clients"
                ],
                "summary": "Delete Client",
                "operationId": "ApiV1DeleteClient",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Delete Client",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
        "/leads/{id}": {
            "get": {
                "tags": [
                    "Leads"
                ],
                "summary": "Get leads information",
                "description": "Returns leads data",
                "operationId": "getIndexById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Leads"
                ],
                "summary": "Update new lead",
                "parameters": [
                    {
                        "name": "lead",
                        "in": "body",
                        "description": "Lead object that needs to be added to the store",
                        "schema": {
                            "properties": {
                                "id": {
                                    "description": "Class LeadsController",
                                    "type": "integer"
                                },
                                "title": {
                                    "description": "Class LeadsController",
                                    "type": "string"
                                },
                                "description": {
                                    "description": "Class LeadsController",
                                    "type": "text"
                                },
                                "status": {
                                    "description": "Class LeadsController",
                                    "type": "available"
                                },
                                "user_assigned_id": {
                                    "description": "Class LeadsController",
                                    "type": "integer"
                                },
                                "user_created_id": {
                                    "description": "Class LeadsController",
                                    "type": "integer"
                                },
                                "client_id": {
                                    "description": "Class LeadsController",
                                    "type": "integer"
                                },
                                "contact_date": {
                                    "description": "Class LeadsController",
                                    "type": "date"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Leads"
                ],
                "summary": "Delete Lead",
                "operationId": "ApiV1DeleteLead",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Delete Lead",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
        "/rules/{id}": {
            "get": {
                "tags": [
                    "Rules"
                ],
                "summary": "Get rules information",
                "description": "Returns rules data",
                "operationId": "getIndexById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Rules"
                ],
                "summary": "Update new rule",
                "parameters": [
                    {
                        "name": "rule",
                        "in": "body",
                        "description": "Rule object that needs to be added to the store",
                        "schema": {
                            "properties": {
                                "id": {
                                    "description": "Class RulesController",
                                    "type": "integer"
                                },
                                "name": {
                                    "description": "Class RulesController",
                                    "type": "string"
                                },
                                "guard_name": {
                                    "description": "Class RulesController",
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/rules/add_permissions/{id}": {
            "get": {
                "tags": [
                    "Rules"
                ],
                "summary": "Set permissions to roles",
                "description": "Returns rules data",
                "operationId": "getIndexById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
        "/sprints/{id}": {
            "get": {
                "tags": [
                    "Sprints"
                ],
                "summary": "Get sprints information",
                "description": "Returns sprints data",
                "operationId": "getIndexById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Sprints"
                ],
                "summary": "Update new sprint",
                "parameters": [
                    {
                        "name": "sprint",
                        "in": "body",
                        "description": "Sprint object that needs to be added to the store",
                        "schema": {
                            "properties": {
                                "id": {
                                    "description": "Class ClientsController",
                                    "type": "integer"
                                },
                                "title": {
                                    "description": "Class ClientsController",
                                    "type": "string"
                                },
                                "description": {
                                    "description": "Class ClientsController",
                                    "type": "text"
                                },
                                "status": {
                                    "description": "Class ClientsController",
                                    "type": "available"
                                },
                                "user_assigned_id": {
                                    "description": "Class ClientsController",
                                    "type": "integer"
                                },
                                "user_created_id": {
                                    "description": "Class ClientsController",
                                    "type": "integer"
                                },
                                "deadline": {
                                    "description": "Class ClientsController",
                                    "type": "date"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Sprints"
                ],
                "summary": "Delete Sprint",
                "operationId": "ApiV1DeleteSprint",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Delete Sprint",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/pointses/{id}": {
            "get": {
                "tags": [
                    "Sprints"
                ],
                "summary": "Get story points",
                "description": "Returns sprints data",
                "operationId": "getIndexById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
        "/voter/{id}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get votes information",
                "description": "Returns tasks data",
                "operationId": "getIndexById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}